fix: avoid unnecessary ipv6
`/etc/hosts` on MacOs has two lines for the entry `localhost`:

```
$ grep localhost /etc/hosts
127.0.0.1	localhost
::1             localhost
```

The app, when given an url, such as `http://localhost:8080`, the
dns resolver will try both of these entries, as specified in the happy
eyeballs rfc (rfc8305).

Since the services in question for the scope of this project only speaks
ipv4, and these are intended to be used only for services running
locally, it's safe to replace 'localhost' with '127.0.0.1'.